# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Pizza Restaurant API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MenuItem {
  id: number;
  name: string;
  description: string;
  price: number;
  category: string;
  /** @nullable */
  imageUrl: string | null;
  isAvailable: boolean;
  isFeatured: boolean;
}

export interface MenuCategory {
  name: string;
  count: number;
}

export interface OrderLineItem {
  menuItemId: number;
  name: string;
  price: number;
  quantity: number;
}

export type OrderInputOrderType = typeof OrderInputOrderType[keyof typeof OrderInputOrderType];


export const OrderInputOrderType = {
  delivery: 'delivery',
  pickup: 'pickup',
} as const;

export interface OrderInput {
  /** @minLength 1 */
  customerName: string;
  customerEmail: string;
  customerPhone: string;
  deliveryAddress?: string;
  orderType: OrderInputOrderType;
  items: OrderLineItem[];
  /** @nullable */
  specialInstructions?: string | null;
}

export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export const OrderStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  preparing: 'preparing',
  ready: 'ready',
  delivered: 'delivered',
  cancelled: 'cancelled',
} as const;

export interface Order {
  id: number;
  customerName: string;
  customerEmail: string;
  customerPhone: string;
  /** @nullable */
  deliveryAddress: string | null;
  orderType: string;
  items: OrderLineItem[];
  total: number;
  status: OrderStatus;
  /** @nullable */
  specialInstructions: string | null;
  createdAt: string;
}

export interface ContactInput {
  /** @minLength 1 */
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @minLength 1 */
  message: string;
}

export interface ContactSubmission {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  phone: string | null;
  message: string;
  createdAt: string;
}

export interface GalleryItem {
  id: number;
  imageUrl: string;
  alt: string;
  category: string;
}

export type GetMenuItemsParams = {
category?: string;
};