# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Pizza Restaurant API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: menu
    description: Menu item operations
  - name: orders
    description: Order operations
  - name: contact
    description: Contact form operations
  - name: gallery
    description: Gallery operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /menu-items:
    get:
      operationId: getMenuItems
      tags: [menu]
      summary: List all menu items
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of menu items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MenuItem"

  /menu-items/featured:
    get:
      operationId: getFeaturedMenuItems
      tags: [menu]
      summary: Get featured menu items
      responses:
        "200":
          description: Featured menu items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MenuItem"

  /menu-items/{id}:
    get:
      operationId: getMenuItem
      tags: [menu]
      summary: Get a single menu item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Menu item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /menu-categories:
    get:
      operationId: getMenuCategories
      tags: [menu]
      summary: Get all menu categories with item counts
      responses:
        "200":
          description: Categories with counts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MenuCategory"

  /orders:
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /contact:
    post:
      operationId: submitContact
      tags: [contact]
      summary: Submit a contact form message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactInput"
      responses:
        "201":
          description: Contact submission received
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactSubmission"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /gallery:
    get:
      operationId: getGalleryItems
      tags: [gallery]
      summary: List gallery images
      responses:
        "200":
          description: Gallery images
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GalleryItem"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    MenuItem:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        price:
          type: number
        category:
          type: string
        imageUrl:
          type: ["string", "null"]
        isAvailable:
          type: boolean
        isFeatured:
          type: boolean
      required:
        - id
        - name
        - description
        - price
        - category
        - imageUrl
        - isAvailable
        - isFeatured

    MenuCategory:
      type: object
      properties:
        name:
          type: string
        count:
          type: integer
      required:
        - name
        - count

    OrderLineItem:
      type: object
      properties:
        menuItemId:
          type: integer
        name:
          type: string
        price:
          type: number
        quantity:
          type: integer
      required:
        - menuItemId
        - name
        - price
        - quantity

    OrderInput:
      type: object
      properties:
        customerName:
          type: string
          minLength: 1
        customerEmail:
          type: string
        customerPhone:
          type: string
        deliveryAddress:
          type: string
        orderType:
          type: string
          enum: [delivery, pickup]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderLineItem"
        specialInstructions:
          type: ["string", "null"]
      required:
        - customerName
        - customerEmail
        - customerPhone
        - orderType
        - items

    Order:
      type: object
      properties:
        id:
          type: integer
        customerName:
          type: string
        customerEmail:
          type: string
        customerPhone:
          type: string
        deliveryAddress:
          type: ["string", "null"]
        orderType:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderLineItem"
        total:
          type: number
        status:
          type: string
          enum: [pending, confirmed, preparing, ready, delivered, cancelled]
        specialInstructions:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - customerName
        - customerEmail
        - customerPhone
        - deliveryAddress
        - orderType
        - items
        - total
        - status
        - specialInstructions
        - createdAt

    ContactInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
        phone:
          type: ["string", "null"]
        message:
          type: string
          minLength: 1
      required:
        - name
        - email
        - message

    ContactSubmission:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        message:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - phone
        - message
        - createdAt

    GalleryItem:
      type: object
      properties:
        id:
          type: integer
        imageUrl:
          type: string
        alt:
          type: string
        category:
          type: string
      required:
        - id
        - imageUrl
        - alt
        - category