# lib/api-zod/src/generated/types/contactInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Pizza Restaurant API
 * OpenAPI spec version: 0.1.0
 */

export interface ContactInput {
  /** @minLength 1 */
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @minLength 1 */
  message: string;
}