# lib/api-zod/src/generated/types/contactSubmission.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Pizza Restaurant API
 * OpenAPI spec version: 0.1.0
 */

export interface ContactSubmission {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  phone: string | null;
  message: string;
  createdAt: Date;
}