# lib/api-zod/src/generated/types/galleryItem.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Pizza Restaurant API
 * OpenAPI spec version: 0.1.0
 */

export interface GalleryItem {
  id: number;
  imageUrl: string;
  alt: string;
  category: string;
}