# lib/api-zod/src/generated/types/menuItem.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Pizza Restaurant API
 * OpenAPI spec version: 0.1.0
 */

export interface MenuItem {
  id: number;
  name: string;
  description: string;
  price: number;
  category: string;
  /** @nullable */
  imageUrl: string | null;
  isAvailable: boolean;
  isFeatured: boolean;
}