# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Pizza Restaurant API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderLineItem } from './orderLineItem';
import type { OrderStatus } from './orderStatus';

export interface Order {
  id: number;
  customerName: string;
  customerEmail: string;
  customerPhone: string;
  /** @nullable */
  deliveryAddress: string | null;
  orderType: string;
  items: OrderLineItem[];
  total: number;
  status: OrderStatus;
  /** @nullable */
  specialInstructions: string | null;
  createdAt: Date;
}