# lib/api-zod/src/generated/types/orderInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Pizza Restaurant API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderInputOrderType } from './orderInputOrderType';
import type { OrderLineItem } from './orderLineItem';

export interface OrderInput {
  /** @minLength 1 */
  customerName: string;
  customerEmail: string;
  customerPhone: string;
  deliveryAddress?: string;
  orderType: OrderInputOrderType;
  items: OrderLineItem[];
  /** @nullable */
  specialInstructions?: string | null;
}