# lib/api-zod/src/generated/types/orderLineItem.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Pizza Restaurant API
 * OpenAPI spec version: 0.1.0
 */

export interface OrderLineItem {
  menuItemId: number;
  name: string;
  price: number;
  quantity: number;
}